+++ src/app/CHANGELOG.md
@@ -7,10 +7,11 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased] - 2025-12-04
 
 ### Added
-- Added toast notifications
-- Added Wish Edit functionality
+- Delete confirmation toast
+- Edit confirmation toast
+- 
 
 
 
 ### Fixed
--
+-